Move controllers to correct namespace
